# apps/backend-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.APIResponse:
    properties:
      count:
        type: integer
      data: {}
      error:
        type: string
      has_more:
        type: boolean
      next_cursor:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    name: gomo6 Team
    url: https://gomo6.wtf
  description: gomo6 — open social platform API. Create bots, integrations, and apps.
  license:
    name: MIT
  termsOfService: https://gomo6.wtf/terms
  title: gomo6 API
  version: "1.0"
paths:
...
  /profiles:
    get:
      description: Get user profiles with optional filters
      parameters:
      - description: Filter by user ID (eq.uuid or in.(uuid,...))
        in: query
        name: id
        type: string
      - description: Filter by username
        in: query
        name: username
        type: string
      - description: Filter by domain
        in: query
        name: domain
        type: string
      - default: 50
        description: Max results (1-100)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: List profiles
      tags:
      - Profiles
  /profiles/{id}:
    get:
      description: Get a user profile by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get profile
      tags:
      - Profiles
    put:
      consumes:
      - application/json
      description: Update user profile (own profile only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - Profiles
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Bearer token for API authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"